move sidebar to ts
